requirements.txt
allauth-django==0.42.6
ArangoDjango==0.0.25
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-bootstrap-modal-forms==2.0.1
django-crispy==1.8.1
django-import-export==2.4.0
django-modelforms==0.3.0
django-utils-six==2.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
xlrd==1.2.0
xlwt==1.3.0

